+++ src/python/pants/backend/python/util_rules/lockfile_test.py
@@ -1,12 +1,19 @@
 # Copyright 2023 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+from __future__ import annotations
+
 from textwrap import dedent
 from typing import Iterable, Type
+from pants.backend.python.util_rules.interpreter_constraints import InterpreterConstraints
+from pants.engine.rules import QueryRule
+from pants.util.ordered_set import FrozenOrderedSet
 
 import pytest
 
 from pants.backend.python.dependency_inference.rules import import_rules
 from pants.backend.python.goals import lockfile
+from pants.backend.python.goals.lockfile import GeneratePythonLockfile
 from pants.backend.python.subsystems.python_tool_base import LockfileRules, PythonToolBase
 from pants.backend.python.target_types import ConsoleScript
 from pants.core.goals import generate_lockfiles
@@ -31,7 +38,7 @@ def _get_generated_lockfile_sentinel(
     ).union_member
 
 
-class FakeTool(PythonToolBase):
+class FakeToolWithSimpleLocking(PythonToolBase):
     options_scope = "cowsay"
     name = "Cowsay"
     help = "A tool to test pants"
@@ -61,12 +68,14 @@ class MockTarget(Target):
 
 @pytest.fixture
 def rule_runner() -> RuleRunner:
+    lockfile_sentinel = _get_generated_lockfile_sentinel(FakeToolWithSimpleLocking.rules(), FakeToolWithSimpleLocking)
     rule_runner = RuleRunner(
         rules=[
             *lockfile.rules(),
             *generate_lockfiles.rules(),
             *import_rules(),
-            *FakeTool.rules(),
+            *FakeToolWithSimpleLocking.rules(),
+            QueryRule(GeneratePythonLockfile, [lockfile_sentinel]),
         ],
         target_types=[MockTarget],
     )
@@ -93,9 +102,45 @@ def test_simple_python_lockfile(rule_runner):
         dedent(
             f"""\
              //   "generated_with_requirements": [
-             //     "{FakeTool.default_version}"
+             //     "{FakeToolWithSimpleLocking.default_version}"
              //   ],
              """
         )
         in lockfile_content
     )
+
+def test_setup_lockfile(rule_runner) -> None:
+    global_constraint = "CPython<4,>=3.8"
+
+    lockfile_sentinel = _get_generated_lockfile_sentinel(FakeToolWithSimpleLocking.rules(), FakeToolWithSimpleLocking)
+
+    def assert_lockfile_request(
+        build_file: str,
+        expected_ics: list[str],
+        *,
+        extra_expected_requirements: list[str] | None = None,
+        extra_args: list[str] | None = None,
+    ) -> None:
+        rule_runner.write_files({"project/BUILD": build_file, "project/f.py": ""})
+        rule_runner.set_options(
+            ["--cowsay-lockfile=lockfile.txt", *(extra_args or [])],
+            env={"PANTS_PYTHON_INTERPRETER_CONSTRAINTS": f"['{global_constraint}']"},
+            env_inherit={"PATH", "PYENV_ROOT", "HOME"},
+        )
+        lockfile_request = rule_runner.request(GeneratePythonLockfile, [lockfile_sentinel()])
+        assert lockfile_request.interpreter_constraints == InterpreterConstraints(expected_ics)
+        assert lockfile_request.requirements == FrozenOrderedSet(
+            [
+                FakeToolWithSimpleLocking.default_version,
+                *FakeToolWithSimpleLocking.default_extra_requirements,
+                *(extra_expected_requirements or ()),
+            ]
+        )
+
+    assert_lockfile_request("python_sources()", FakeToolWithSimpleLocking.default_interpreter_constraints)
+    assert_lockfile_request("target()", FakeToolWithSimpleLocking.default_interpreter_constraints)
+    # Since the SIMPLE locking mechanism doesn't look at ICs, this will still use tool ICs.
+    assert_lockfile_request(
+        "python_sources(interpreter_constraints=['CPython<4,>=3.7'])", FakeToolWithSimpleLocking.default_interpreter_constraints
+    )
+
